Projeto funcional com requisitos desejados, pendente impl. "testes"

+++ src/main/java/battlemovies/modelo/Ranking.java
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Ranking {
     private String nome;
-    private Long pontuacao;
+    private int pontuacao;
 
     public Ranking(String txtLinhas) {
         String[] split = txtLinhas.split(",");
         this.nome = split[0].trim().replace("\"", "");
-        this.pontuacao= Long.parseLong(split[1].trim().replace("\"", ""));
+        this.pontuacao= Integer.parseInt(split[1].trim().replace("\"", ""));
     }
 }
